refactor: rename import  "../auth/auth" to "../auth"
